+++ ruitk/elements/abuse.js
@@ -1,0 +1,5 @@
+import { Ruitk } from "./elements";
+
+Ruitk.prototype.abuse = function () {
+    
+}