+++ doc/changelog.md
@@ -1,5 +1,7 @@
 
 ## 0.9.11
+#### documentation Changes
+ - Added a documentation page for security issues
 #### Fallen Changes
  - Added new `markdown` element in `staging.js`
  - Added some error checking to the `generate` function